requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0
joblib==1.3.2
gunicorn==21.2.0
pytest==7.4.0
pytest-mock==3.11.1
